rename CookieManager.js -> ../AuthManager.js
